Game Switcher - General Improvements
* Moved most game-switcher logic from existing ES code base into self-contained module.

* Add missing include to SDL_timer.h to fix build error.

* Added new setting and support to boot straight into game-switcher if quick-resume not in-progress.

---------
