+++ es-app/src/guis/GuiMenu.cpp
@@ -3926,7 +3926,7 @@ void GuiMenu::openUISettings(bool selectGameSwitcherEnable)
 	// Game Switcher Settings submenu - only visible when enabled
 	if (baseGameSwitcherEnabled)
 	{
-		s->addEntry(_("GAME SWITCHER SETTINGS"), true, [this]() { openGameSwitcherSettings(); });
+		s->addEntry(_("GAME SWITCHER SETTINGS"), true, [window]() { GuiGameSwitcher::openSettings(window); });
 	}
 
 	// Dynamic menu recreation when toggle changes
@@ -3999,101 +3999,7 @@ void GuiMenu::openUISettings(bool selectGameSwitcherEnable)
 	mWindow->pushGui(s);
 }
 
-void GuiMenu::openGameSwitcherSettings(bool selectMarqueeEnable, bool selectPlayInfoEnable)
-{
-	bool baseMarqueeEnabled = Settings::getInstance()->getBool("GameSwitcherMarqueeEnabled");
-	bool basePlayInfoEnabled = Settings::getInstance()->getBool("GameSwitcherPlayInfoEnabled");
 
-	auto s = new GuiSettings(mWindow, _("GAME SWITCHER SETTINGS").c_str());
-
-	// Game Switcher count setting
-	auto gameSwitcherCount = std::make_shared<SliderComponent>(mWindow, 5.f, 25.f, 1.f, "");
-	gameSwitcherCount->setValue((float)Settings::getInstance()->getInt("GameSwitcherCount"));
-	s->addWithLabel(_("AVAILABLE SAVE COUNT"), gameSwitcherCount);
-	s->addSaveFunc([gameSwitcherCount] {
-		Settings::getInstance()->setInt("GameSwitcherCount", (int)Math::round(gameSwitcherCount->getValue()));
-	});
-
-	// Game Switcher animation speed setting
-	auto gameSwitcherSpeed = std::make_shared<SliderComponent>(mWindow, 100.f, 1000.f, 50.f, "ms");
-	gameSwitcherSpeed->setValue((float)Settings::getInstance()->getInt("GameSwitcherAnimationSpeed"));
-	s->addWithLabel(_("ANIMATION SPEED (MS)"), gameSwitcherSpeed);
-	s->addSaveFunc([gameSwitcherSpeed] {
-		Settings::getInstance()->setInt("GameSwitcherAnimationSpeed", (int)Math::round(gameSwitcherSpeed->getValue()));
-	});
-
-	// Enable Marquee toggle
-	auto marqueeEnable = std::make_shared<SwitchComponent>(mWindow);
-	marqueeEnable->setState(baseMarqueeEnabled);
-	s->addWithLabel(_("ENABLE MARQUEE"), marqueeEnable, selectMarqueeEnable);
-	s->addSaveFunc([marqueeEnable] {
-		Settings::getInstance()->setBool("GameSwitcherMarqueeEnabled", marqueeEnable->getState());
-	});
-
-	// Game Switcher marquee size setting - only shown when marquee is enabled
-	if (baseMarqueeEnabled)
-	{
-		auto gameSwitcherMarquee = std::make_shared<SliderComponent>(mWindow, 20.f, 80.f, 5.f, "%");
-		gameSwitcherMarquee->setValue((float)Settings::getInstance()->getInt("GameSwitcherMarqueeSize"));
-		s->addWithLabel(_("MARQUEE SIZE"), gameSwitcherMarquee);
-		s->addSaveFunc([gameSwitcherMarquee] {
-			Settings::getInstance()->setInt("GameSwitcherMarqueeSize", (int)Math::round(gameSwitcherMarquee->getValue()));
-		});
-	}
-
-	// Dynamic menu recreation when marquee toggle changes
-	marqueeEnable->setOnChangedCallback([this, s, baseMarqueeEnabled, marqueeEnable]()
-	{
-		bool enabled = marqueeEnable->getState();
-		if (baseMarqueeEnabled != enabled)
-		{
-			Settings::getInstance()->setBool("GameSwitcherMarqueeEnabled", enabled);
-			delete s;
-			openGameSwitcherSettings(true, false);
-		}
-	});
-
-	// Enable Play Information toggle
-	auto playInfoEnable = std::make_shared<SwitchComponent>(mWindow);
-	playInfoEnable->setState(basePlayInfoEnabled);
-	s->addWithLabel(_("ENABLE PLAY INFORMATION"), playInfoEnable, selectPlayInfoEnable);
-	s->addSaveFunc([playInfoEnable] {
-		Settings::getInstance()->setBool("GameSwitcherPlayInfoEnabled", playInfoEnable->getState());
-	});
-
-	// Game Switcher info background opacity setting - only shown when play info is enabled
-	if (basePlayInfoEnabled)
-	{
-		auto gameSwitcherBgOpacity = std::make_shared<SliderComponent>(mWindow, 0.f, 100.f, 5.f, "%");
-		gameSwitcherBgOpacity->setValue((float)Settings::getInstance()->getInt("GameSwitcherInfoBackgroundOpacity"));
-		s->addWithLabel(_("BACKGROUND OPACITY"), gameSwitcherBgOpacity);
-		s->addSaveFunc([gameSwitcherBgOpacity] {
-			Settings::getInstance()->setInt("GameSwitcherInfoBackgroundOpacity", (int)Math::round(gameSwitcherBgOpacity->getValue()));
-		});
-	}
-
-	// Enable Launch Animation toggle
-	auto launchAnimEnable = std::make_shared<SwitchComponent>(mWindow);
-	launchAnimEnable->setState(Settings::getInstance()->getBool("GameSwitcherLaunchAnimationEnabled"));
-	s->addWithLabel(_("ENABLE LAUNCH ANIMATION"), launchAnimEnable);
-	s->addSaveFunc([launchAnimEnable] {
-		Settings::getInstance()->setBool("GameSwitcherLaunchAnimationEnabled", launchAnimEnable->getState());
-	});
-
-	// Dynamic menu recreation when play info toggle changes
-	playInfoEnable->setOnChangedCallback([this, s, basePlayInfoEnabled, playInfoEnable]()
-	{
-		bool enabled = playInfoEnable->getState();
-		if (basePlayInfoEnabled != enabled)
-		{
-			Settings::getInstance()->setBool("GameSwitcherPlayInfoEnabled", enabled);
-			delete s;
-			openGameSwitcherSettings(false, true);
-		}
-	});
-
-	mWindow->pushGui(s);
-}
 
 void GuiMenu::openSoundSettings()
 {
