+++ es-app/src/guis/GuiMenu.h
@@ -50,7 +50,6 @@ private:
         void openScreensaverOptions();  
         void openSoundSettings();
         void openUISettings(bool selectGameSwitcherEnable = false);
-        void openGameSwitcherSettings(bool selectMarqueeEnable = false, bool selectPlayInfoEnable = false);
         void openUpdatesSettings();
         
         void openSystemSettings();
